Move mapping functions and main render function in render package.
