+++ report/topology.go
@@ -2,7 +2,6 @@ package report
 
 import (
 	"fmt"
-	"log"
 	"net"
 	"reflect"
 	"strings"
@@ -75,110 +74,6 @@ func NewTopology() Topology {
 		EdgeMetadatas: map[string]EdgeMetadata{},
 		NodeMetadatas: map[string]NodeMetadata{},
 	}
-}
-
-// RenderBy transforms a given Topology into a set of RenderableNodes, which
-// the UI will render collectively as a graph. Note that a RenderableNode will
-// always be rendered with other nodes, and therefore contains limited detail.
-//
-// RenderBy takes a a MapFunc, which defines how to group and label nodes. Npdes
-// with the same mapped IDs will be merged.
-func (t Topology) RenderBy(mapFunc MapFunc, pseudoFunc PseudoFunc) RenderableNodes {
-	nodes := RenderableNodes{}
-
-	// Build a set of RenderableNodes for all non-pseudo probes, and an
-	// addressID to nodeID lookup map. Multiple addressIDs can map to the same
-	// RenderableNodes.
-	var (
-		source2mapped = map[string]string{} // source node ID -> mapped node ID
-		source2host   = map[string]string{} // source node ID -> origin host ID
-	)
-	for nodeID, metadata := range t.NodeMetadatas {
-		mapped, ok := mapFunc(metadata)
-		if !ok {
-			continue
-		}
-
-		// mapped.ID needs not be unique over all addressIDs. If not, we merge with
-		// the existing data, on the assumption that the MapFunc returns the same
-		// data.
-		existing, ok := nodes[mapped.ID]
-		if ok {
-			mapped.Merge(existing)
-		}
-
-		mapped.Origins = mapped.Origins.Add(nodeID)
-		nodes[mapped.ID] = mapped
-		source2mapped[nodeID] = mapped.ID
-		source2host[nodeID] = metadata[HostNodeID]
-	}
-
-	// Walk the graph and make connections.
-	for src, dsts := range t.Adjacency {
-		var (
-			srcNodeID, ok = ParseAdjacencyID(src)
-			//srcOriginHostID, _, ok2 = ParseNodeID(srcNodeID)
-			srcHostNodeID     = source2host[srcNodeID]
-			srcRenderableID   = source2mapped[srcNodeID] // must exist
-			srcRenderableNode = nodes[srcRenderableID]   // must exist
-		)
-		if !ok {
-			log.Printf("bad adjacency ID %q", src)
-			continue
-		}
-
-		for _, dstNodeID := range dsts {
-			dstRenderableID, ok := source2mapped[dstNodeID]
-			if !ok {
-				pseudoNode, ok := pseudoFunc(srcNodeID, srcRenderableNode, dstNodeID)
-				if !ok {
-					continue
-				}
-				dstRenderableID = pseudoNode.ID
-				nodes[dstRenderableID] = pseudoNode
-				source2mapped[dstNodeID] = dstRenderableID
-			}
-
-			srcRenderableNode.Adjacency = srcRenderableNode.Adjacency.Add(dstRenderableID)
-			srcRenderableNode.Origins = srcRenderableNode.Origins.Add(srcHostNodeID)
-			srcRenderableNode.Origins = srcRenderableNode.Origins.Add(srcNodeID)
-			edgeID := MakeEdgeID(srcNodeID, dstNodeID)
-			if md, ok := t.EdgeMetadatas[edgeID]; ok {
-				srcRenderableNode.Metadata.Merge(md.Transform())
-			}
-		}
-
-		nodes[srcRenderableID] = srcRenderableNode
-	}
-
-	return nodes
-}
-
-// EdgeMetadata gives the metadata of an edge from the perspective of the
-// srcRenderableID. Since an edgeID can have multiple edges on the address
-// level, it uses the supplied mapping function to translate address IDs to
-// renderable node (mapped) IDs.
-func (t Topology) EdgeMetadata(mapFunc MapFunc, srcRenderableID, dstRenderableID string) EdgeMetadata {
-	metadata := EdgeMetadata{}
-	for edgeID, edgeMeta := range t.EdgeMetadatas {
-		src, dst, ok := ParseEdgeID(edgeID)
-		if !ok {
-			log.Printf("bad edge ID %q", edgeID)
-			continue
-		}
-		if src != TheInternet {
-			mapped, _ := mapFunc(t.NodeMetadatas[src])
-			src = mapped.ID
-		}
-		if dst != TheInternet {
-			mapped, _ := mapFunc(t.NodeMetadatas[dst])
-			dst = mapped.ID
-		}
-		if src == srcRenderableID && dst == dstRenderableID {
-			metadata.Flatten(edgeMeta)
-		}
-	}
-	return metadata
 }
 
 // Squash squashes all non-local nodes in the topology to a super-node called
